% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.models.R
\name{compare.models}
\alias{compare.models}
\title{WAIC-based model comparison}
\usage{
compare.models(...)
}
\arguments{
\item{...}{MCMC output from either \code{\link{nimbleMCMC}} or \code{\link{runMCMC}} functions in the \code{nimble} R package. Note that in argument \code{WAIC} should be set to TRUE.}
}
\value{
A table containing WAIC, delta WAIC, and WAIC weights.
}
\description{
Compute delta WAIC and WAIC weights for model comparison.
}
