% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dAsymLaplace.R
\name{dAsymLaplace}
\alias{dAsymLaplace}
\alias{dAsymLaplace,}
\alias{rAsymLaplace}
\title{Asymmetric Laplace Distribution}
\usage{
dAsymLaplace(x, mu, sigma, tau, log)

rAsymLaplace(n, mu, sigma, tau)
}
\arguments{
\item{x}{value to be computed.}

\item{mu}{location parameter.}

\item{sigma}{scale parameter.}

\item{tau}{asymmetry parameter.}

\item{log}{TRUE or 1 to return log probability. FALSE or 0 to return probability.}

\item{n}{number of random draws. Currently only n = 1 is supported, but the argument exists for standardization of "r" functions.}
}
\description{
Density and random generation for a Asymmetric Laplace Distribution.
}
\author{
Enrico Crema
}
