% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shcal20}
\alias{shcal20}
\title{IntCal20 radiocarbon age calibration curve for the Southern hemisphere.}
\format{
A data.frame with the following fields:
\describe{
\item{\code{CalBP}}{ID of each radiocarbon date}
\item{\code{C14Age}}{Radiocarbon age in 14C years BP}
\item{\code{C14Age.sigma}}{Radiocarbon age error}
\item{\code{Delta14C}}{Labcode of the radiocarbon date}
\item{\code{Delta14C.sigma}}{Material of the dated sample}
}
}
\source{
https://intcal.org/curves/shcal20.14c
}
\usage{
shcal20
}
\description{
IntCal20 radiocarbon age calibration curve for the Southern hemisphere.
}
\references{
Reimer, P. J., Austin, W. E. N., Bard, E., Bayliss, A., Blackwell, P. G., Ramsey, C. B., et al. (2020). The IntCal20 Northern Hemisphere Radiocarbon Age Calibration Curve (0–55 cal kBP). Radiocarbon, 62(4), 725–757. https://doi.org/10.1017/RDC.2020.41
}
\keyword{datasets}
